ファイルを分割、fetchのレスポンスを各ステータスコードのUnionにした
